requirements.txt
charset-normalizer==3.1.0
idna==3.4
meapi==0.7.0
python-decouple==3.8
requests==2.28.2
urllib3==1.26.15
pyTelegramBotAPI==4.10.0